web-animations/interfaces/Animation/effect.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Animation.effect</title>
<link rel="help" href="http://w3c.github.io/web-animations/#dom-animation-effect">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
<body>
<script>
'use strict';

test(function(t) {
  var div = createDiv(t);
  var animation = div.animate({left: ['100px', '100px']}, {fill: 'forwards'});
  var effect = animation.effect;

  assert_equals(getComputedStyle(div).left, '100px', 'animation is initially having an effect');

  animation.effect = null;
  assert_equals(getComputedStyle(div).left, 'auto', 'animation no longer has an effect');

  animation.effect = effect;
  assert_equals(getComputedStyle(div).left, '100px', 'animation has an effect again');
}, 'Clearing and setting Animation.effect should update the effect of the animation');

</script>
</body>
